requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
flask
pypdf
pydantic==1.10.13
docarray==0.32.1
python-dotenv
-e . #this is for setup.py to set up local variable.

